requirements.txt
boto3==1.17.22
botocore==1.20.22
jmespath==0.10.0
python-dateutil==2.8.1
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.3
psycopg2-binary==2.8.6
#python-dotenv
#serverless-dot-env